% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_places.R
\name{cdc_places}
\alias{cdc_places}
\title{CDC PLACES geographic data on health-related burden and outcome}
\source{
\url{https://www.cdc.gov/places/}
}
\usage{
cdc_places()
}
\description{
PLACES, a collaboration between CDC, the Robert Wood Johnson Foundation, and
the CDC Foundation, allows local health departments and jurisdictions
regardless of population size and urban-rural status to better understand the
burden and geographic distribution of health-related outcomes in their areas
and assist them in planning public health interventions.
}
\details{
PLACES provides model-based population-level analysis and community estimates
to all counties, places (incorporated and census designated places), census
tracts, and ZIP Code Tabulation Areas (ZCTAs) across the United States.

PLACES is an extension of the original 500 Cities Project that provided city
and census tract estimates for chronic disease risk factors, health outcomes,
and clinical preventive services use for the 500 largest US cities.

They were accessed on September 25, 2021.
}
\section{Download}{
 -
\url{https://github.com/seandavi/UCCCReporter/raw/main/inst/extdata/PLACES__Local_Data_for_Better_Health__ZCTA_Data_2020_release.csv.gz}
}

\section{Geographic level}{

\itemize{
\item zipcode
}
}

\examples{
places <- cdc_places()

head(places)

# All available measures:
unique(places$Measure)


}
\seealso{
Other Data resources: 
\code{\link{cdc_social_vulnerability_index}()},
\code{\link{cell_line_ancestry}()},
\code{\link{census_tract_zip_crosswalk}()},
\code{\link{rural_zips}()},
\code{\link{zipcode_shapes}()}
}
\concept{Data resources}
