% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_social_vulnerability_index.R
\name{cdc_social_vulnerability_index}
\alias{cdc_social_vulnerability_index}
\title{US CDC Social Vulnerability Index}
\source{
\url{https://svi.cdc.gov/}
}
\usage{
cdc_social_vulnerability_index()
}
\value{
a data.frame
}
\description{
The CDC's Social Vulnerability Index (SVI), created and maintained
by the Geospatial Research, Analysis, and Services Program (GRASP),
uses US Census data to determine the social vulnerability of every
county and tract. This index ranks each county and tract based upon
15 social factors including poverty, lack of vehicle access, and
crowded housing, and groups them into four related themes: 1)
Socioeconomic, 2) Housing Composition and Disability, 3) Minority
Status and Language, and 4) Housing and Transportation.
}
\details{
Theme rankings: For each of the four themes, we summed the
percentiles for the variables comprising each theme. We ordered the
summed percentiles for each theme to determine theme-specific
percentile rankings. The four summary theme ranking variables,
detailed in the Data Dictionary below, are:
\itemize{
\item Socioeconomic - RPL_THEME1
\item Household Composition & Disability - RPL_THEME2
\item Minority Status & Language - RPL_THEME3
\item Housing Type & Transportation - RPL_THEME4
}

Overall tract rankings: We summed the sums for each theme, ordered
the tracts, and then calculated overall percentile rankings. Please
note; taking the sum of the sums for each theme is the same as
summing individual variable rankings. The overall tract summary
ranking variable is RPL_THEM

For detailed documentation, see
\url{https://svi.cdc.gov/Documents/Data/2018_SVI_Data/SVI2018Documentation.pdf}
}
\section{Download}{
 -
\url{https://github.com/seandavi/UCCCReporter/raw/main/inst/extdata/SVI2018_US.csv.gz}
}

\section{Geographic level}{

\itemize{
\item census tract
}
}

\examples{

res = cdc_social_vulnerability_index()

# column names
colnames(res)

}
\seealso{
Other Data resources: 
\code{\link{cdc_places}()},
\code{\link{census_tract_zip_crosswalk}()},
\code{\link{rural_zips}()},
\code{\link{zipcode_shapes}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{Data resources}
