% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_line_ancestry.R
\name{cell_line_ancestry}
\alias{cell_line_ancestry}
\title{Cell line ancestry for common cancer cell lines}
\source{
\url{http://ecla.moffitt.org/}
}
\usage{
cell_line_ancestry()
}
\description{
Data are stored at extdata/cancer-cell-lines-ancestry.csv.gz
They were accessed on November 15, 2021.
}
\section{Download}{

\itemize{
\item \url{https://github.com/seandavi/UCCCReporter/raw/main/inst/extdata/cancer-cell-lines-ancestry.csv.gz}
}
}

\examples{
cl_ancestry = cell_line_ancestry()

head(cl_ancestry)

# Source of cell line
table(cl_ancestry$source)

# cell lines in CCLE?
table(cl_ancestry$ccle_membership)

table(cl_ancestry$ethnicity)

anc_mat = as.matrix(cl_ancestry[,12:18])
heatmap(anc_mat,margins = c(9,1), labRow=NA)

}
\references{
Dutil J, Chen Z, Monteiro AN, Teer JK, Eschrich SA. An Interactive Resource to Probe Genetic Diversity and Estimated Ancestry in Cancer Cell Lines. Cancer Res. 2019 Apr 1;79(7):1263-1273. doi: 10.1158/0008-5472.CAN-18-2747. Epub 2019 Mar 20. PMID: 30894373; PMCID: PMC6445675.
\itemize{
\item \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6445675/}
}
}
\seealso{
Other Data resources: 
\code{\link{cdc_places}()},
\code{\link{cdc_social_vulnerability_index}()},
\code{\link{census_tract_zip_crosswalk}()},
\code{\link{rural_zips}()},
\code{\link{zipcode_shapes}()}
}
\concept{Data resources}
