% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censustract_zip_crosswalk.R
\name{census_tract_zip_crosswalk}
\alias{census_tract_zip_crosswalk}
\title{data.table mapping zips and census tracts}
\source{
\url{https://www.huduser.gov/portal/datasets/usps_crosswalk.html}
}
\usage{
census_tract_zip_crosswalk()
}
\description{
Data are stored at extdata/ZIP_TRACT_062021.xlsx for convenience.
They were accesses on September 25, 2021.
}
\section{Geographic level}{

\itemize{
\item zipcode
\item census tract
}
}

\section{Download}{

\itemize{
\item \url{https://github.com/seandavi/UCCCReporter/raw/main/inst/extdata/TRACT_ZIP_092021.csv.gz}
}
}

\examples{
ctract_zip = census_tract_zip_crosswalk()

head(ctract_zip)

}
\seealso{
Other Data resources: 
\code{\link{cdc_places}()},
\code{\link{cdc_social_vulnerability_index}()},
\code{\link{rural_zips}()},
\code{\link{zipcode_shapes}()}
}
\concept{Data resources}
