% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_patient_report.R
\name{get_patient_report}
\alias{get_patient_report}
\title{get demographic report from Bigquery or cache}
\usage{
get_patient_report(use_cache = TRUE)
}
\arguments{
\item{use_cache}{logical(1) whether or not to use
values from cache. If cache does not exist, data are downloaded.}
}
\description{
get demographic report from Bigquery or cache
}
\seealso{
Other Internal utilities: 
\code{\link{set_gcs_auth_file}()}
}
\concept{Internal utilities}
