% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rural_zips.R
\name{rural_zips}
\alias{rural_zips}
\title{data.table of rural zipcodes from HRSA}
\source{
\url{https://www.hrsa.gov/rural-health/about-us/definition/datafiles.html}
}
\usage{
rural_zips()
}
\description{
These data were compiled by Federal Office of Rural Health Policy (FORHP)
at the Health Resources and Services Administration (HRSA). This data
resources gives a simple list of zipcodes that are considered \emph{rural}.
}
\details{
They were accessed on September 25, 2021.
}
\section{Download}{

\itemize{
\item \url{https://github.com/seandavi/UCCCReporter/raw/main/inst/extdata/forhp-eligible-zips.xlsx}
}
}

\section{Geographic level}{

\itemize{
\item zipcode
\item state
}
}

\examples{
rural_zip_data = rural_zips()

head(rural_zip_data)

# top states by number of rural counties
rural_zip_data[,.(.N),by=state][order(N,decreasing=TRUE)][1:10,]

}
\seealso{
Other Data resources: 
\code{\link{cdc_places}()},
\code{\link{cdc_social_vulnerability_index}()},
\code{\link{census_tract_zip_crosswalk}()},
\code{\link{zipcode_shapes}()}
}
\concept{Data resources}
