% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_auth_file.R
\name{set_gcs_auth_file}
\alias{set_gcs_auth_file}
\title{set auth token path}
\usage{
set_gcs_auth_file(GCS_AUTH_FILE)
}
\arguments{
\item{GCS_AUTH_FILE}{character(1) path to GCS json file}
}
\description{
set auth token path
}
\seealso{
Other Internal utilities: 
\code{\link{get_patient_report}()}
}
\concept{Internal utilities}
