% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcode_shapes.R
\name{zipcode_shapes}
\alias{zipcode_shapes}
\title{Colorado shape files from census bureau}
\usage{
zipcode_shapes(
  state = "colorado",
  year = 2010,
  ...,
  progress_bar = interactive()
)
}
\arguments{
\item{state}{character(1) default 'colorado', passed directly to \code{tigris::zctas()}}

\item{year}{integer(1) default 2010, passed directly to \code{tigris::zctas()}}

\item{...}{passed directly to \code{tigris::zctas()}}
}
\description{
This function simply wraps \code{tigris::zctas()} to grab the zip code shape
files for later plotting.
}
\section{Geographic level}{

\itemize{
\item zipcode
}
}

\examples{

zipshapes = zipcode_shapes()

tmap::tmap_mode('plot')

tmap::tm_shape(zipshapes) +
  tmap::tm_fill('ALAND10') +
  tmap::tm_borders(col="gray",alpha=0.5) +
  tmap::tm_basemap("Esri.WorldStreetMap")


}
\seealso{
Other Data resources: 
\code{\link{cdc_places}()},
\code{\link{cdc_social_vulnerability_index}()},
\code{\link{census_tract_zip_crosswalk}()},
\code{\link{rural_zips}()}
}
\concept{Data resources}
